+++ packages/runtime/src/index.ts
@@ -7,4 +7,5 @@ export * from "./dockerode-adapter.js";
 export * from "./factory.js";
 export * from "./server-runtime.js";
 export * from "./host-capabilities.js";
+export * from "./udp-listen.js";
 export * from "./steamcmd.js";